+++ contracts/ownership/DelayedClaimable.sol
@@ -11,8 +11,8 @@ import './Claimable.sol';
  */
 contract DelayedClaimable is Claimable {
 
-  uint public end;
-  uint public start;
+  uint256 public end;
+  uint256 public start;
 
   /**
    * @dev Used to specify the time period during which a pending 
@@ -20,7 +20,7 @@ contract DelayedClaimable is Claimable {
    * @param _start The earliest time ownership can be claimed.
    * @param _end The latest time ownership can be claimed. 
    */
-  function setLimits(uint _start, uint _end) onlyOwner {
+  function setLimits(uint256 _start, uint256 _end) onlyOwner {
     if (_start > _end)
         throw;
     end = _end;
